Refactor subject banners Rewrite the classifier subject banners as pure functions which take a subject as a prop. Remove `storiesOf` and `withKnobs` from the component stories. Rewrite stories as CSF.
