+++ packages/lib-classifier/src/components/Classifier/components/Banners/components/AlreadySeenBanner/index.js
@@ -1,1 +1,1 @@
-export { default } from './AlreadySeenBannerContainer'
+export { default } from './AlreadySeenBanner'
