+++ packages/lib-classifier/src/components/Classifier/components/Banners/components/RetiredBanner/index.js
@@ -1,1 +1,1 @@
-export { default } from './RetiredBannerContainer'
+export { default } from './RetiredBanner'
