+++ packages/lib-classifier/src/components/Classifier/components/Banners/components/UserHasFinishedWorkflowBanner/UserHasFinishedWorkflowBanner.spec.js
@@ -1,15 +1,15 @@
 import { shallow } from 'enzyme'
 import React from 'react'
 
-import RetiredBannerContainer from './RetiredBannerContainer'
+import UserHasFinishedWorkflowBanner from './UserHasFinishedWorkflowBanner'
 import Banner from '../Banner'
 
 let wrapper
 let componentWrapper
 
-describe('Component > RetiredBannerContainer', function () {
+describe('Component > UserHasFinishedWorkflowBanner', function () {
   before(function () {
-    wrapper = shallow(<RetiredBannerContainer.wrappedComponent />)
+    wrapper = shallow(<UserHasFinishedWorkflowBanner />)
     componentWrapper = wrapper.find(Banner)
   })
 
@@ -22,28 +22,28 @@ describe('Component > RetiredBannerContainer', function () {
   })
 
   it('should pass a `background` prop to <Banner />', function () {
-    expect(componentWrapper.prop('background')).to.equal('status-critical')
+    expect(componentWrapper.prop('background')).to.equal('status-warning')
   })
 
   it('should pass a `bannerText` prop to <Banner />', function () {
-    expect(componentWrapper.prop('bannerText')).to.equal('Finished: This subject has been retired')
+    expect(componentWrapper.prop('bannerText')).to.equal('Finished: You\'ve seen everything')
   })
 
   it('should pass a `tooltipText` prop to <Banner />', function () {
     const expectedText = [
-      'This subject already has enough classifications, so yours won\'t be used in its analysis.',
+      'You\'ve already classified everything in this workflow, so further classifications on this subject won\'t be used in its analysis.',
       'If you\'re looking to help, try choosing a different workflow or contributing to a different project.'
     ]
     expect(componentWrapper.prop('tooltipText')).to.deep.equal(expectedText)
   })
 
   describe('when the banner should show', function () {
-    it('should show if the subject has been retired and the workflow isn\'t finished', function () {
+    it('should show if the user has finished the workflow', function () {
       wrapper.setProps({
         subject: {
           finished_workflow: false,
           id: '1',
-          retired: true
+          user_has_finished_workflow: true
         }
       })
       expect(wrapper.find(Banner).prop('show')).to.be.true()
@@ -56,46 +56,21 @@ describe('Component > RetiredBannerContainer', function () {
       expect(wrapper.find(Banner).prop('show')).to.be.false()
     })
 
-    it('shouldn\'t show when subject hasn\'t been retired', function () {
+    it('shouldn\'t show when the user hasn\'t finished the workflow', function () {
       wrapper.setProps({
         subject: {
-          finished_workflow: false,
           id: '1',
-          retired: false
+          user_has_finished_workflow: false
         }
       })
       expect(wrapper.find(Banner).prop('show')).to.be.false()
     })
 
-    it('shouldn\'t show if the subject has been retired and the workflow is finished', function () {
+    it('shouldn\'t show when the workflow is finished', function () {
       wrapper.setProps({
         subject: {
           finished_workflow: true,
           id: '1',
-          retired: true
-        }
-      })
-      expect(wrapper.find(Banner).prop('show')).to.be.false()
-    })
-
-    it('shouldn\'t show the banner if the subject has been retired, but the workflow is finished', function () {
-      wrapper.setProps({
-        subject: {
-          already_seen: true,
-          finished_workflow: true,
-          id: '1',
-          retired: true
-        }
-      })
-      expect(wrapper.find(Banner).prop('show')).to.be.false()
-    })
-
-    it('shouldn\'t show the banner if the subject has been retired, but the user has finished the workflow', function () {
-      wrapper.setProps({
-        subject: {
-          already_seen: true,
-          id: '1',
-          retired: true,
           user_has_finished_workflow: true
         }
       })
