+++ packages/lib-classifier/src/components/Classifier/components/Banners/components/UserHasFinishedWorkflowBanner/index.js
@@ -1,1 +1,1 @@
-export { default } from './UserHasFinishedWorkflowBannerContainer'
+export { default } from './UserHasFinishedWorkflowBanner'
