+++ packages/lib-classifier/src/components/Classifier/components/Banners/components/WorkflowIsFinishedBanner/index.js
@@ -1,1 +1,1 @@
-export { default } from './WorkflowIsFinishedBannerContainer'
+export { default } from './WorkflowIsFinishedBanner'
